Progress on #71
