+++ src/main/java/frc/robot/subsystems/arm/ArmConstants.java
@@ -1,0 +1,24 @@
+package frc.robot.subsystems.arm;
+
+import edu.wpi.first.math.util.Units;
+import edu.wpi.first.wpilibj.simulation.SingleJointedArmSim;
+
+public final class ArmConstants {
+  public static final int ARM_MOTOR_ID = 44;
+  public static final double ARM_MOTOR_KP = 5; // 3.5
+  public static final double ARM_MOTOR_KI = 0.00; // 0.01
+  public static final double ARM_MOTOR_KD = 0.0;
+  public static final double ARM_MOTOR_KG = 0.0;
+  public static final double ARM_MOTOR_KV = 0.00;
+  public static final double ARM_MOTOR_KFF = 0.0;
+  public static final double ARM_MOTOR_KS = 0.0; // 0.1
+  public static final int ARM_MOTOR_CURRENT_LIMIT = 60;
+  // Sim constants
+  public static final double GEARING = 1.0 / 1.125;
+  public static final double LENGTH = Units.inchesToMeters(22.9);
+  public static final double MOI =
+      SingleJointedArmSim.estimateMOI(ArmConstants.LENGTH, Units.lbsToKilograms(5.3754589));
+  public static final double MIN_ANGLE = Units.degreesToRadians(-90);
+  public static final double MAX_ANGLE = Units.degreesToRadians(103);
+  public static final double STARTING_ANGLE = Units.degreesToRadians(-90);
+}
