+++ src/main/java/frc/robot/subsystems/climb/ClimbIOTalonFX.java
@@ -12,7 +12,6 @@ import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.units.measure.AngularVelocity;
 import edu.wpi.first.units.measure.Current;
 import edu.wpi.first.units.measure.Voltage;
-import frc.robot.Constants.ClimbConstants;
 import frc.robot.generated.TunerConstants;
 
 public class ClimbIOTalonFX implements ClimbIO {
