+++ src/main/java/frc/robot/subsystems/elevator/Elevator.java
@@ -8,7 +8,6 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.RunCommand;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
-import frc.robot.Constants.ElevatorConstants;
 import frc.robot.Constants.Mode;
 import java.util.function.DoubleSupplier;
 import org.littletonrobotics.junction.Logger;
