+++ src/main/java/frc/robot/subsystems/endeffector/EndEffectorConstants.java
@@ -1,0 +1,18 @@
+package frc.robot.subsystems.endeffector;
+
+public final class EndEffectorConstants {
+  public static final int EFFECTORID = 43;
+  public static final int EFFECTOR_CURRENT_LIMIT = 20;
+  public static final double EFFECTOR_KP = 0.1;
+  public static final double EFFECTOR_KI = 0.0;
+  public static final double EFFECTOR_KD = 0.0;
+  public static final double EFFECTOR_KS = 0.0;
+  public static final double EFFECTOR_KV = 0.0;
+  public static final double EFFECTOR_KFF = 0.0;
+  public static final double EFFECTOR_KG = 0.05;
+  public static final int LASERCAN_1ID = 51;
+  public static final int LASERCAN_2ID = 52;
+  // Sim Constants
+  public static final double MOI = 0.0002048478;
+  public static final double GEARING = 1;
+}
