+++ src/main/java/frc/robot/subsystems/endeffector/EndEffectorIOTalonFX.java
@@ -12,7 +12,6 @@ import edu.wpi.first.math.util.Units;
 import edu.wpi.first.units.measure.AngularVelocity;
 import edu.wpi.first.units.measure.Current;
 import edu.wpi.first.units.measure.Voltage;
-import frc.robot.Constants.EndEffectorConstants;
 
 public class EndEffectorIOTalonFX implements EndEffectorIO {
