+++ src/main/java/frc/robot/subsystems/groundintake/GroundIntakeIO.java
@@ -6,7 +6,6 @@ import com.revrobotics.spark.SparkLowLevel.MotorType;
 import com.revrobotics.spark.config.ClosedLoopConfig.FeedbackSensor;
 import com.revrobotics.spark.config.SparkBaseConfig.IdleMode;
 import com.revrobotics.spark.config.SparkFlexConfig;
-import frc.robot.Constants.GroundIntakeConstants;
 import org.littletonrobotics.junction.AutoLog;
 
 public interface GroundIntakeIO {
