+++ src/main/java/frc/robot/subsystems/led/AddressableLEDIO.java
@@ -3,7 +3,6 @@ package frc.robot.subsystems.led;
 import edu.wpi.first.wpilibj.AddressableLED;
 import edu.wpi.first.wpilibj.AddressableLEDBuffer;
 import edu.wpi.first.wpilibj.LEDPattern;
-import frc.robot.Constants.LEDConstants;
 import frc.robot.subsystems.led.LEDIO.LedIOInputs;
 
 public class AddressableLEDIO implements LEDIO {
