+++ src/main/java/frc/robot/subsystems/led/LEDConstants.java
@@ -1,0 +1,12 @@
+package frc.robot.subsystems.led;
+
+import static edu.wpi.first.units.Units.Meters;
+
+import edu.wpi.first.units.measure.Distance;
+
+public final class LEDConstants {
+  public static final int LED_PWM_PORT = 0;
+  public static final int LED_LENGTH = 123;
+  public static final int BARGE_RANGE = 1;
+  public static final Distance LED_SPACING = Meters.of(1.0 / 50.0);
+}
