+++ src/main/java/frc/robot/subsystems/vision/VisionIOPhotonVisionSim.java
@@ -1,10 +1,9 @@
 package frc.robot.subsystems.vision;
 
-import static frc.robot.Constants.VisionConstants.aprilTagLayout;
+import static frc.robot.subsystems.vision.VisionConstants.aprilTagLayout;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Transform3d;
-import frc.robot.Constants.VisionConstants;
 import frc.robot.subsystems.vision.VisionIO.VisionIOInputs;
 import java.util.function.Supplier;
 import org.photonvision.simulation.PhotonCameraSim;
